removed fuction definition from .hpp filr
